+++ src/navigation/ModalStack.tsx
@@ -2,7 +2,6 @@ import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 import MakeCouponModal from '../components/Campagin/MakeCouponModal';
 import MakePinPointModal from '../components/Campagin/MakePinPointModal';
-import MakePinPointQuiz from '../components/Campagin/MakePinPointQuiz';
 
 interface Props {
 
@@ -14,8 +13,6 @@ const ModalStack = (props: Props) => {
     return (
         <Stack.Navigator>
             <Stack.Screen name="MakePinPointModal" component={MakePinPointModal} />
-            <Stack.Screen name="MakePinPointQuiz" component={MakePinPointQuiz} />
-
             <Stack.Screen name="MakeCouponModal" component={MakeCouponModal} />
         </Stack.Navigator>
     )
